refactor brats part
